renamed download module for unittest.mock to work with it. added commented out mock test of upload/create_md5 tht requires python 3.6+
