+++ src/download/__init__.py
@@ -1,1 +1,1 @@
-from .download import download
+from .download_module import download
